librbd/cache/pwl/ssd: make log entry pointers 64 bit (on-disk format change)
Fixes: https://tracker.ceph.com/issues/50675

(cherry picked from commit c091ec3471973f41717b237d500b8e4374af660f)

+++ src/librbd/cache/pwl/LogEntry.h
@@ -25,7 +25,7 @@ class GenericLogEntry {
 public:
   WriteLogCacheEntry ram_entry;
   WriteLogCacheEntry *cache_entry = nullptr;
-  uint32_t log_entry_index = 0;
+  uint64_t log_entry_index = 0;
   bool completed = false;
   GenericLogEntry(uint64_t image_offset_bytes = 0, uint64_t write_bytes = 0)
     : ram_entry(image_offset_bytes, write_bytes) {
